+++ param.sh
@@ -15,10 +15,14 @@ source /var/lib/boot2docker/extension/environment.sh
 # -------------------------------------------------------------------
 # Change param values if possible
 # -------------------------------------------------------------------
-BOOT2DOCKER_EXTENSION_PARAM_FILE="$BOOT2DOCKER_EXTENSION_PARAM_DIR/$1"
+BOOT2DOCKER_EXTENSION_PARAM_FILE="${BOOT2DOCKER_EXTENSION_PARAM_DIR}/${1}"
 if [ -f $BOOT2DOCKER_EXTENSION_PARAM_FILE ]; then
-    echo "$2" > $BOOT2DOCKER_EXTENSION_PARAM_FILE
-    echo "[INFO] $BOOT2DOCKER_EXTENSION_PARAM_FILE updated with value=$2"
+    if [[ "${3}" == "--append" ]]; then
+        echo "${2}" >> $BOOT2DOCKER_EXTENSION_PARAM_FILE
+    else
+        echo "${2}" > $BOOT2DOCKER_EXTENSION_PARAM_FILE
+    fi
+    echo "[INFO] ${BOOT2DOCKER_EXTENSION_PARAM_FILE} updated with value=${2}"
 else
-    echo "[WARN] $BOOT2DOCKER_EXTENSION_PARAM_FILE not found so not updated"
+    echo "[WARN] ${BOOT2DOCKER_EXTENSION_PARAM_FILE} not found so not updated"
 fi