+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## 17.09.0-ce (2017-10-10)
 - FEATURE: now scripts called in vagrant file are stored here in `vagrant` folder for remote call
+- FEATURE: now `param.sh` can take a third optional argument `--append` to append new param a the end of the file
 
 ## 17.09.0-ce (2017-10-09)
 - UPDATE: now `version_check` won't not occur in background anymore (avoid keypress need)